front-Tabel

+++ BackStore/controllers/buy.controller.js
@@ -2,10 +2,10 @@ import buyProductModel from "../models/buyModel.model.js";
 
 export const buyProduct = async (req, res) => {
   const productId = req.params.product;
-  const userId = req.user.userId;
+  const userId = req.user.id;
   const product = await buyProductModel.findOne({ productId, userId });
   if (product) {
-    await buyProductModel.findOneAndDelete({ productId });
+    await buyProductModel.findOneAndDelete({ productId,userId });
     return res.send({
       msg: "Product removed from buy list",
       isAdded: false,
@@ -19,7 +19,7 @@ export const buyProduct = async (req, res) => {
 };
 
 export const getBuyProduct = async (req, res) => {
-  const userId = req.res.id;
-  const buyProduct = await buyProductModel.find(userId).populate("productId");
+  const userId = req.user.id;
+  const buyProduct = await buyProductModel.find({userId}).populate("productId");
   res.send(buyProduct.map((item) => item.productId));
 };
